moved macros
