+++ src/main/kotlin/com/tezos/lang/michelson/parser/macro/ConditionalMacroMetadata.kt
@@ -1,10 +1,12 @@
-package com.tezos.lang.michelson.parser
+package com.tezos.lang.michelson.parser.macro
 
 import com.tezos.lang.michelson.psi.PsiAnnotationType
 
-class CompareMacroMetadata : MacroMetadata {
+class ConditionalMacroMetadata : MacroMetadata {
     companion object {
-        val NAMES = setOf("CMPEQ", "CMPNEQ", "CMPLT", "CMPGT", "CMPLE", "CMPGE")
+        val NAMES = setOf("IFEQ", "IFNEQ", "IFLT", "IFGT", "IFLE", "IFGE",
+                "IFCMPEQ", "IFCMPNEQ", "IFCMPLT", "IFCMPGT", "IFCMPLE", "IFCMPGE",
+                "IF_SOME")
     }
 
     override fun validate(macro: String): Pair<String, Int>? {
@@ -15,7 +17,7 @@ class CompareMacroMetadata : MacroMetadata {
         }
     }
 
-    override fun requiredBlocks(): Int = 0
+    override fun requiredBlocks(): Int = 2
 
     override fun supportedAnnotations(type: PsiAnnotationType, macro: String): Int = 0
 }
