+++ src/main/kotlin/com/tezos/lang/michelson/parser/macro/DipMacroMetadata.kt
@@ -1,7 +1,6 @@
-package com.tezos.lang.michelson.parser
+package com.tezos.lang.michelson.parser.macro
 
 import com.tezos.lang.michelson.psi.PsiAnnotationType
-import java.util.*
 import java.util.regex.Pattern
 
 /**
