+++ src/main/kotlin/com/tezos/lang/michelson/parser/macro/SetCadrMacroMetadata.kt
@@ -1,20 +1,23 @@
-package com.tezos.lang.michelson.parser
+package com.tezos.lang.michelson.parser.macro
 
 import com.tezos.lang.michelson.psi.PsiAnnotationType
 import java.util.regex.Pattern
 
 /**
- * Supports the dynamic CADR macro, as defined at http://tezos.gitlab.io/betanet/whitedoc/michelson.html#syntactic-conveniences.
- * CADR A syntactic sugar for accessing fields in nested pairs.
+ * Supports the dynamic SET_CADR macro, as defined at http://tezos.gitlab.io/betanet/whitedoc/michelson.html#syntactic-conveniences.
+ * SET_CADR is syntactic sugar for setting fields in nested pairs.
+ *
+ * The variable annotation on SET_C\[AD]+R and MAP_C\[AD]+R annotates the resulting toplevel pair
+ * while its field annotation is used to check that the modified field is the expected one.
  */
-class CadrMacroMetadata : MacroMetadata {
+class SetCadrMacroMetadata : MacroMetadata {
     private companion object {
-        val regexp = Pattern.compile("C[AD]+R")
+        val regexp = Pattern.compile("SET_C[AD]+R")
     }
 
     override fun validate(macro: String): Pair<String, Int>? {
         if (!regexp.matcher(macro).matches()) {
-            return "Invalid CADR macro. It must match 'C[AD]+R'. " to 0
+            return "Invalid SET_CADR macro. It must match 'SET_C[AD]+R'. " to 0
         }
 
         // validation against the current stack isn't possible atm, so there's no further validation to do here
