+++ src/test/kotlin/com/tezos/lang/michelson/parser/CompareMacroMetadataTest.kt
@@ -1,5 +1,6 @@
 package com.tezos.lang.michelson.parser
 
+import com.tezos.lang.michelson.parser.macro.CompareMacroMetadata
 import com.tezos.lang.michelson.psi.PsiAnnotationType
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
