+++ src/test/kotlin/com/tezos/lang/michelson/parser/MapCadrMacroMetadataTest.kt
@@ -1,5 +1,6 @@
 package com.tezos.lang.michelson.parser
 
+import com.tezos.lang.michelson.parser.macro.MapCadrMacroMetadata
 import com.tezos.lang.michelson.psi.PsiAnnotationType
 import org.junit.Assert.*
 import org.junit.Test
