+++ src/test/kotlin/com/tezos/lang/michelson/parser/PairMacroMetadataTest.kt
@@ -1,5 +1,6 @@
 package com.tezos.lang.michelson.parser
 
+import com.tezos.lang.michelson.parser.macro.PairMacroMetadata
 import org.junit.Assert.assertEquals
 import org.junit.Test
